Improve thread panic and assertion failure patterns
The
    %E  left:%m,
    %C right:%m %f:%l:%c
    %Z
set of patterns was pretty frustrating. The quickfix list entry had just
the LHS value as the message, and was lacking %f:%l%c to allow you to
jump to the failure. The RHS pattern currently (with Rust 1.74) doesn't
print the %f:%l:%c location, so it wasn't being included in the entry.

This is all replaced by a better rule

    %+E.*panicked at %f:%l:%c:
    %+Cassertion.*

Which creates a _much_ better error message for :clist

* For panics not triggered by assertions, at least there's still an
  error message of "... panicked ..."
* For panics that are assertions, the kind of assertion (assert_eq! or
  assert!) is printed, but not the assertion values (because that can
  make the error message become very unwieldy)

This can be tested by saving the contents of a Cargo run (either build
or test) to a file

    # Use .rs so that the Rust errorformat rules get used
    $ cargo build |& tee -a build-output.rs
    $ cargo test |& tee -a test-output.rs
    $ vim build-output.rs
    " This is the magic. Load the quickfix list from the current buffer.
    :cgetbuffer
    :clist
    :copen