Fix ugly Error highlighting on fenced code blocks in attributes
The following code

    #[cfg_attr(feature = "doc-diagrams", doc = svgbobdoc::transform!(
    /// ```svgbob
    /// ...
    /// ```
    ))]

had the whole doc comment highlighted as rustCommentLineDocError, which
was quite ugly.

I don't think this is the correct fix for the issue, but I really
struggled to understand and extend the rustComment.*RustCode regions to
work inside the rustAttributeContents cluster.

This at least makes it less ugly for now.